Create tabbed_parallel_corpus_maker.py
Once you have all your verses (or other line numbering in another kind of textual corpus) exactly exactly parallel as separate text files, with a newline (\n) separating each verse or line, use this program to turn the two files into a single parallel corpus file with each source-target pair separated by a tab. This is the format required for the translator in this repository.